added hero images need to format
