+++ blog/pasta-travels/index.mdx
@@ -1,6 +1,10 @@
 ---
 title: "Travel for Pasta"
 date: "2021-11-10"
+hero_image: "./man-with-pasta-basket.jpg"
+hero_image_alt: "A guy holding a big basket full of cooked pasta "
+hero_image_credit_text: "Daniel Watson"
+hero_image_credit_link: "https://unsplash.com/photos/yhX9Ntr9HjM"
 ---
 
 ### Discover Pasta Shapes in Asia
