+++ blog/sec-post.mdx
@@ -1,6 +1,10 @@
 ---
 title: "Second Post"
 date: "2021-04-24"
+hero_image: "./christopher-ayme-ocZ-_Y7-Ptg-unsplash.jpg"
+hero_image_alt: "A gray pitbull relaxing on the sidewalk with its tongue hanging out"
+hero_image_credit_text: "Christopher Ayme"
+hero_image_credit_link: "https://unsplash.com/photos/ocZ-_Y7-Ptg"
 ---
 
 Here's another post! It's even better than the first one!
