+++ blog/tks-pasta-recipes/index.mdx
@@ -1,6 +1,10 @@
 ---
 title: "Thanksgiving Pasta Recipes"
 date: "2021-11-27"
+hero_image: "./pumpkins-hay.jpg"
+hero_image_alt: "A field with pumpkins on hay"
+hero_image_credit_text: "Joseph Gonzalez"
+hero_image_credit_link: "https://unsplash.com/photos/E-EEQSjDdck"
 ---
 
 ### Popular pasta dishes for Thanksgiving
