+++ blog/unique-recipes/index.mdx
@@ -1,6 +1,10 @@
 ---
 title: "Unique Recipes"
 date: "2021-08-26"
+hero_image: "./linguini-hand.jpg"
+hero_image_alt: "A hand with flour holding linguini pasta"
+hero_image_credit_text: "Jorge Zapata"
+hero_image_credit_link: "https://unsplash.com/photos/4nXkhLCrkLo"
 ---
 
 ### Unique Pasta Recipes
