+++ blog/weird-shapes/index.mdx
@@ -1,6 +1,10 @@
 ---
 title: "Weird Pasta Shapes"
 date: "2021-06-20"
+hero_image: "./spaghetti-paella.jpg"
+hero_image_alt: "Paella with spaghetti pasta"
+hero_image_credit_text: "Olayinka Babalola"
+hero_image_credit_link: "https://unsplash.com/photos/r01ZopTiEV8"
 ---
 
 ### Weird Shapes!
