--- src/pages/blog.js
@@ -1,39 +1,0 @@
-import * as React from "react";
-import { graphql } from "gatsby";
-import { MDXRenderer } from "gatsby-plugin-mdx";
-
-const BlogPage = ({ data }) => {
-  return (
-    <div>
-      {data.allMdx.nodes.map(node => {
-        return (
-          <article key={node.id}>
-            <h2>{node.frontmatter.title}</h2>
-            <p>Posted: {node.frontmatter.date}</p>
-            <MDXRenderer>{node.body}</MDXRenderer>
-          </article>
-        );
-      })}
-    </div>
-  );
-};
-
-export const query = graphql`
-  query {
-    allMdx(
-      sort: { fields: frontmatter___date, order: DESC }
-      filter: { frontmatter: { date: { ne: null } } }
-    ) {
-      nodes {
-        frontmatter {
-          title
-          date(formatString: "MMMM DD, YYYY")
-        }
-        id
-        body
-      }
-    }
-  }
-`;
-
-export default BlogPage;
